Add CollaboratorPermissionsEditor component
