+++ src/client/components/EditUserForm/CollaboratorPermissions/CollaboratorPermissions.tsx
@@ -1,16 +1,15 @@
 import './CollaboratorPermissions.scss'
-import React, { useState } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { Collaborator, CollaboratorPermissions as CollabPermissions } from 'meta/user'
 
 import { useAppDispatch } from 'client/store'
-import { useSections } from 'client/store/metadata'
 import { UserManagementActions } from 'client/store/ui/userManagement'
 import { useCountryIso } from 'client/hooks'
 import { useCycleRouteParams } from 'client/hooks/useRouteParams'
 import CollaboratorAccessList from 'client/components/CollaboratorAccessList'
-import CollaboratorAccessModal from 'client/components/CollaboratorAccessModal'
+import CollaboratorPermissionsEditor from 'client/components/user/CollaboratorPermissionsEditor'
 
 type Props = {
   userRole: Collaborator
@@ -23,19 +22,6 @@ const CollaboratorPermissions = (props: Props) => {
 
   const { assessmentName, cycleName } = useCycleRouteParams()
   const countryIso = useCountryIso()
-  const sections = useSections()
-
-  const [modalOptions, setModalOptions] = useState<{ open: boolean }>({ open: false })
-
-  const _onEditPermissionsClick = () => {
-    setModalOptions({ open: true })
-  }
-
-  const _onEditPermissionsClose = () => {
-    setModalOptions({ open: false })
-  }
-
-  if (!sections) return null
 
   const permissions = (userRole.permissions as CollabPermissions) || undefined
 
@@ -60,17 +46,8 @@ const CollaboratorPermissions = (props: Props) => {
       <div className="edit-user__form-field edit-user__form-field-premissions">
         <CollaboratorAccessList permissions={permissions} />
 
-        <button className="btn-xs btn-primary" onClick={_onEditPermissionsClick} type="button">
-          {t('userManagement.editPermissions')}
-        </button>
+        <CollaboratorPermissionsEditor permissions={permissions} onPermissionsChange={handlePermissionsChange} />
       </div>
-
-      <CollaboratorAccessModal
-        open={modalOptions.open}
-        permissions={permissions}
-        onClose={_onEditPermissionsClose}
-        onPermissionsChange={handlePermissionsChange}
-      />
     </div>
   )
 }
