+++ src/client/components/user/CollaboratorPermissionsEditor/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CollaboratorPermissionsEditor'
